# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /surveys:
    get:
      operationId: listSurveys
      tags: [surveys]
      summary: List all surveys
      responses:
        "200":
          description: List of surveys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Survey"
    post:
      operationId: createSurvey
      tags: [surveys]
      summary: Create a new survey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyBody"
      responses:
        "201":
          description: Created survey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
  /surveys/{surveyId}:
    get:
      operationId: getSurvey
      tags: [surveys]
      summary: Get a survey with its questions
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survey with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyWithQuestions"
        "404":
          description: Survey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSurvey
      tags: [surveys]
      summary: Update a survey
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSurveyBody"
      responses:
        "200":
          description: Updated survey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
    delete:
      operationId: deleteSurvey
      tags: [surveys]
      summary: Delete a survey
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /surveys/{surveyId}/responses:
    post:
      operationId: submitResponse
      tags: [responses]
      summary: Submit a response to a survey
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResponseBody"
      responses:
        "201":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseRecord"
        "404":
          description: Survey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /surveys/{surveyId}/results:
    get:
      operationId: getSurveyResults
      tags: [results]
      summary: Get aggregated results for a survey
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Aggregated results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurveyResults"
        "404":
          description: Survey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /surveys/{surveyId}/responses/raw:
    get:
      operationId: getRawResponses
      tags: [results]
      summary: Get all raw responses for a survey
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Raw response list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResponseWithAnswers"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Survey:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        responseCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, isActive, responseCount, createdAt, updatedAt]
    SurveyWithQuestions:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        responseCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required: [id, title, isActive, responseCount, createdAt, updatedAt, questions]
    Question:
      type: object
      properties:
        id:
          type: integer
        surveyId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [text, multiple_choice, rating, yes_no]
        options:
          type: array
          items:
            type: string
        isRequired:
          type: boolean
        orderIndex:
          type: integer
        section:
          type: string
        sectionDescription:
          type: string
      required: [id, surveyId, text, type, options, isRequired, orderIndex]
    CreateSurveyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestionBody"
      required: [title, questions]
    CreateQuestionBody:
      type: object
      properties:
        text:
          type: string
        type:
          type: string
          enum: [text, multiple_choice, rating, yes_no]
        options:
          type: array
          items:
            type: string
        isRequired:
          type: boolean
        orderIndex:
          type: integer
      required: [text, type]
    UpdateSurveyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        questions:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestionBody"
    SubmitResponseBody:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
      required: [answers]
    AnswerInput:
      type: object
      properties:
        questionId:
          type: integer
        value:
          type: string
        values:
          type: array
          items:
            type: string
      required: [questionId]
    ResponseRecord:
      type: object
      properties:
        id:
          type: integer
        surveyId:
          type: integer
        userId:
          type: string
        submittedAt:
          type: string
      required: [id, surveyId, submittedAt]
    ResponseWithAnswers:
      type: object
      properties:
        id:
          type: integer
        surveyId:
          type: integer
        submittedAt:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
      required: [id, surveyId, submittedAt, answers]
    Answer:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        questionText:
          type: string
        value:
          type: string
        values:
          type: array
          items:
            type: string
      required: [id, questionId, questionText]
    SurveyResults:
      type: object
      properties:
        surveyId:
          type: integer
        surveyTitle:
          type: string
        totalResponses:
          type: integer
        completionRate:
          type: number
        questionResults:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
        recentResponses:
          type: array
          items:
            $ref: "#/components/schemas/ResponseRecord"
      required: [surveyId, surveyTitle, totalResponses, completionRate, questionResults, recentResponses]
    QuestionResult:
      type: object
      properties:
        questionId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
        totalAnswers:
          type: integer
        textAnswers:
          type: array
          items:
            type: string
        choiceCounts:
          type: object
          additionalProperties:
            type: integer
        averageRating:
          type: number
        ratingDistribution:
          type: object
          additionalProperties:
            type: integer
      required: [questionId, questionText, questionType, totalAnswers]
...