# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all surveys
 */
export const ListSurveysResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  isActive: zod.boolean(),
  responseCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListSurveysResponse = zod.array(ListSurveysResponseItem);

/**
 * @summary Create a new survey
 */
export const CreateSurveyBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  questions: zod.array(
    zod.object({
      text: zod.string(),
      type: zod.enum(["text", "multiple_choice", "rating", "yes_no"]),
      options: zod.array(zod.string()).optional(),
      isRequired: zod.boolean().optional(),
      orderIndex: zod.number().optional(),
    }),
  ),
});

/**
 * @summary Get a survey with its questions
 */
export const GetSurveyParams = zod.object({
  surveyId: zod.coerce.number(),
});

export const GetSurveyResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  isActive: zod.boolean(),
  responseCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  questions: zod.array(
    zod.object({
      id: zod.number(),
      surveyId: zod.number(),
      text: zod.string(),
      type: zod.enum(["text", "multiple_choice", "rating", "yes_no"]),
      options: zod.array(zod.string()),
      isRequired: zod.boolean(),
      orderIndex: zod.number(),
      section: zod.string().optional(),
      sectionDescription: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Update a survey
 */
export const UpdateSurveyParams = zod.object({
  surveyId: zod.coerce.number(),
});

export const UpdateSurveyBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  isActive: zod.boolean().optional(),
  questions: zod
    .array(
      zod.object({
        text: zod.string(),
        type: zod.enum(["text", "multiple_choice", "rating", "yes_no"]),
        options: zod.array(zod.string()).optional(),
        isRequired: zod.boolean().optional(),
        orderIndex: zod.number().optional(),
      }),
    )
    .optional(),
});

export const UpdateSurveyResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().optional(),
  isActive: zod.boolean(),
  responseCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a survey
 */
export const DeleteSurveyParams = zod.object({
  surveyId: zod.coerce.number(),
});

/**
 * @summary Submit a response to a survey
 */
export const SubmitResponseParams = zod.object({
  surveyId: zod.coerce.number(),
});

export const SubmitResponseBody = zod.object({
  answers: zod.array(
    zod.object({
      questionId: zod.number(),
      value: zod.string().optional(),
      values: zod.array(zod.string()).optional(),
    }),
  ),
});

/**
 * @summary Get aggregated results for a survey
 */
export const GetSurveyResultsParams = zod.object({
  surveyId: zod.coerce.number(),
});

export const GetSurveyResultsResponse = zod.object({
  surveyId: zod.number(),
  surveyTitle: zod.string(),
  totalResponses: zod.number(),
  completionRate: zod.number(),
  questionResults: zod.array(
    zod.object({
      questionId: zod.number(),
      questionText: zod.string(),
      questionType: zod.string(),
      totalAnswers: zod.number(),
      textAnswers: zod.array(zod.string()).optional(),
      choiceCounts: zod.record(zod.string(), zod.number()).optional(),
      averageRating: zod.number().optional(),
      ratingDistribution: zod.record(zod.string(), zod.number()).optional(),
    }),
  ),
  recentResponses: zod.array(
    zod.object({
      id: zod.number(),
      surveyId: zod.number(),
      userId: zod.string().optional(),
      submittedAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get all raw responses for a survey
 */
export const GetRawResponsesParams = zod.object({
  surveyId: zod.coerce.number(),
});

export const GetRawResponsesResponseItem = zod.object({
  id: zod.number(),
  surveyId: zod.number(),
  submittedAt: zod.string(),
  answers: zod.array(
    zod.object({
      id: zod.number(),
      questionId: zod.number(),
      questionText: zod.string(),
      value: zod.string().optional(),
      values: zod.array(zod.string()).optional(),
    }),
  ),
});
export const GetRawResponsesResponse = zod.array(GetRawResponsesResponseItem);

/**
 * @summary Get overview stats across all surveys
 */
export const GetDashboardOverviewResponse = zod.object({
  totalSurveys: zod.number(),
  activeSurveys: zod.number(),
  totalResponses: zod.number(),
  responsesToday: zod.number(),
  recentActivity: zod.array(
    zod.object({
      surveyId: zod.number(),
      surveyTitle: zod.string(),
      responseCount: zod.number(),
      lastResponseAt: zod.string().optional(),
    }),
  ),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
      isAdmin: zod.boolean(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce
    .string()
    .optional()
    .describe(
      "Relative path to redirect to after login (must start with `\/`). Defaults to `\/`.",
    ),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});